+++ foobarqix/src/main/java/com/sebdev/foobarqix/rule/RuleDivisor.java
@@ -2,20 +2,29 @@ package com.sebdev.foobarqix.rule;
 
 import com.sebdev.foobarqix.FooBarQixEnum;
 
+/**
+ * Rule used to search if a value is divible by a number from FooBarFixEnum
+ */
 public class RuleDivisor implements Rule {
 
     private final int value;
 
+    /**
+     * Prepare the rule for a number to evaluate
+     */
     public RuleDivisor(int value) {
         this.value = value;
     }
 
+    /**
+     * return true if entered number is divisible by the foobarqix number in parameter
+     */
     @Override
     public boolean isValid(FooBarQixEnum fooBarQixEnum) {
-        return isFooBarQixMultiple(fooBarQixEnum);
+        return isDivisibleBy(fooBarQixEnum.getDivisor());
     }
 
-    private boolean isFooBarQixMultiple(FooBarQixEnum fooBarQixEnum) {
-        return (value % fooBarQixEnum.getDivisor()) == 0;
+    private boolean isDivisibleBy(int divisor) {
+        return (value % divisor) == 0;
     }
 }
